Initial commit.  There are some issues to be hammered out.
A few issues:
1. The Motor doesn't keep running when a speed is set.
For eg: If I wrote BrickPi.MotorSpeed[PORT_A]=200; just in the main
part, the motor just jerks once but doesn't keep running.

The only way to keep it running, right now, is making an infinite loop
and keep setting the speed.
It stops running even when delay, sleep or wait function is used. Which
means, there is NO way to keep it running for a specified number of
seconds/ms!

2. The touch sensor doesn't work when battery is plugged in.
I gave detailed info on this issue in the previous mail. There is some
issue in the A2D converter of Arduino.

3. The Raspberry Pi CPU throttles at 100% when a BrickPi code is run.
I don't know if this is due to the infinite loop or some background
processing.
Also, there's no way to run programs on BrickPi without an infinite
loop, which needs to be looked upon. This will seriously affect the
performance and battery life.

4. The Digital/I2C sensors are not working right now on our machine.
But it gives garbage values when a sensor is plugged.